Modifying the construction of the cost matrix
Before the cost matrix was calculate just with the distance between the prediction and the detection. We are multiplying the distance with the cosine between the vector previous center - prediction and the vector previous center - detection. This will allow to discard detections that doesn’t follow the natural trajectory/flow of the vehicle.