Structure modified with abstract class for analysis
Implemented abstract class Analyzer which has some specific/basic methods. This class needs to be extended with a subclass that will perform the desired analysis with the tracked objects: count, track, average time, … Following this structure is easier to write our own class with the integration of it with the rest of the system.
Implemented also a function that acts as a factory. Takes the analysisConfig.json and automaticaly create the class with the configuration specified inside the file.